+++ client/src/pages/automation/projects/ProjectWorkflowListItem.tsx
@@ -1,4 +1,3 @@
-import WorkflowDialog from '@/components/WorkflowDialog/WorkflowDialog';
 import {
     AlertDialog,
     AlertDialogAction,
@@ -25,6 +24,7 @@ import {
     useDuplicateWorkflowMutation,
     useUpdateWorkflowMutation,
 } from '@/mutations/projects.mutations';
+import WorkflowDialog from '@/pages/automation/project/components/WorkflowDialog';
 import {ProjectKeys} from '@/queries/projects.queries';
 import {DotsVerticalIcon} from '@radix-ui/react-icons';
 import {useQueryClient} from '@tanstack/react-query';
